test: moving tests to dedicated repository
IMI-2282
