+++ bin/build.sh
@@ -1,0 +1,3 @@
+#!/bin/sh
+set -exu
+docker-compose build
