+++ bin/wait_then_run
@@ -26,11 +26,9 @@ done
 
 mkdir -p minio/escrow-results
 
-export PYTHONPATH=$(pwd)
+#export PYTHONPATH=$(pwd)
 if [ ! -z $* ]; then
   $*
 else
-  python3 hmt_escrow/job.py
-  python3 hmt_escrow/storage.py
-  python3 hmt_escrow/eth_bridge.py
+  python3 -m unittest discover -s test/hmt_escrow
 fi
