+++ hmt_escrow/job.py
@@ -1,24 +1,20 @@
 #!/usr/bin/env python3
-import os
-import sys
 import logging
-import unittest
-from time import sleep
-from unittest.mock import MagicMock, patch, call
-from functools import partial
+import os
 from decimal import Decimal
 from enum import Enum
 from typing import Dict, List, Tuple, Optional, Any
 
+from basemodels import Manifest
+from eth_keys import keys
+from eth_utils import decode_hex
 from web3 import Web3
 from web3.contract import Contract
 from web3.types import Wei
-from eth_keys import keys
-from eth_utils import decode_hex
 
+from hmt_escrow import utils
 from hmt_escrow.eth_bridge import (
     get_hmtoken,
-    get_contract_interface,
     get_escrow,
     get_factory,
     deploy_factory,
@@ -27,9 +23,7 @@ from hmt_escrow.eth_bridge import (
     Retry,
     HMTOKEN_ADDR,
 )
-from hmt_escrow import utils
 from hmt_escrow.storage import download, upload
-from basemodels import Manifest
 
 GAS_LIMIT = int(os.getenv("GAS_LIMIT", 4712388))
 
@@ -39,10 +33,11 @@ LOG = logging.getLogger("hmt_escrow.job")
 Status = Enum("Status", "Launched Pending Partial Paid Complete Cancelled")
 
 
-def status(escrow_contract: Contract, gas_payer: str, gas: int = GAS_LIMIT) -> Enum:
+def status(escrow_contract: Contract, gas_payer: str,
+           gas: int = GAS_LIMIT) -> Enum:
     """Returns the status of the Job.
 
-    >>> from test_manifest import manifest
+    >>> from test.hmt_escrow.utils import manifest
     >>> credentials = {
     ... 	"gas_payer": "0x1413862C2B7054CDbfdc181B83962CB0FC11fD92",
     ... 	"gas_payer_priv": "28e516f1e2f99e96a48a23cea1f94ee5f073403a1c68e818263f0eb898f1c8e5"
@@ -73,11 +68,11 @@ def status(escrow_contract: Contract, gas_payer: str, gas: int = GAS_LIMIT) -> E
 
 
 def manifest_url(
-    escrow_contract: Contract, gas_payer: str, gas: int = GAS_LIMIT
+        escrow_contract: Contract, gas_payer: str, gas: int = GAS_LIMIT
 ) -> str:
     """Retrieves the deployed manifest url uploaded on Job initialization.
 
-    >>> from test_manifest import manifest
+    >>> from test.hmt_escrow.utils import manifest
     >>> credentials = {
     ... 	"gas_payer": "0x1413862C2B7054CDbfdc181B83962CB0FC11fD92",
     ... 	"gas_payer_priv": "28e516f1e2f99e96a48a23cea1f94ee5f073403a1c68e818263f0eb898f1c8e5"
@@ -106,7 +101,7 @@ def manifest_url(
 
 
 def manifest_hash(
-    escrow_contract: Contract, gas_payer: str, gas: int = GAS_LIMIT
+        escrow_contract: Contract, gas_payer: str, gas: int = GAS_LIMIT
 ) -> str:
     """Retrieves the deployed manifest hash uploaded on Job initialization.
 
@@ -115,7 +110,7 @@ def manifest_hash(
     ... 	"gas_payer_priv": "28e516f1e2f99e96a48a23cea1f94ee5f073403a1c68e818263f0eb898f1c8e5"
     ... }
     >>> rep_oracle_pub_key = b"2dbc2c2c86052702e7c219339514b2e8bd4687ba1236c478ad41b43330b08488c12c8c1797aa181f3a4596a1bd8a0c18344ea44d6655f61fa73e56e743f79e0d"
-    >>> from test_manifest import manifest
+    >>> from test.hmt_escrow.utils import manifest
     >>> job = Job(credentials, manifest)
     >>> job.launch(rep_oracle_pub_key)
     True
@@ -139,14 +134,16 @@ def manifest_hash(
 
 
 def is_trusted_handler(
-    escrow_contract: Contract, handler_addr: str, gas_payer: str, gas: int = GAS_LIMIT
+        escrow_contract: Contract, handler_addr: str, gas_payer: str,
+        gas: int = GAS_LIMIT
 ) -> bool:
     return escrow_contract.functions.areTrustedHandlers(handler_addr).call(
         {"from": gas_payer, "gas": Wei(gas)}
     )
 
 
-def launcher(escrow_contract: Contract, gas_payer: str, gas: int = GAS_LIMIT) -> str:
+def launcher(escrow_contract: Contract, gas_payer: str,
+             gas: int = GAS_LIMIT) -> str:
     """Retrieves the details on what eth wallet launched the job
 
     Args:
@@ -187,13 +184,13 @@ class Job:
     """
 
     def __init__(
-        self,
-        credentials: Dict[str, str],
-        escrow_manifest: Manifest = None,
-        factory_addr: str = None,
-        escrow_addr: str = None,
-        multi_credentials: List[Tuple] = [],
-        retry: Retry = None,
+            self,
+            credentials: Dict[str, str],
+            escrow_manifest: Manifest = None,
+            factory_addr: str = None,
+            escrow_addr: str = None,
+            multi_credentials: List[Tuple] = [],
+            retry: Retry = None,
     ):
         """Initializes a Job instance with values from a Manifest class and
         checks that the provided credentials are valid. An optional factory
@@ -205,7 +202,7 @@ class Job:
         ... 	"gas_payer": "0x1413862C2B7054CDbfdc181B83962CB0FC11fD92",
         ... 	"gas_payer_priv": "28e516f1e2f99e96a48a23cea1f94ee5f073403a1c68e818263f0eb898f1c8e5"
         ... }
-        >>> from test_manifest import manifest
+        >>> from test.hmt_escrow.utils import manifest
         >>> job = Job(credentials, manifest)
         >>> job.gas_payer == credentials["gas_payer"]
         True
@@ -281,15 +278,18 @@ class Job:
             multi_credentials, **credentials
         )
         if not main_credentials_valid:
-            raise ValueError("Given private key doesn't match the ethereum address.")
+            raise ValueError(
+                "Given private key doesn't match the ethereum address.")
 
         self.gas_payer = Web3.toChecksumAddress(credentials["gas_payer"])
         self.gas_payer_priv = credentials["gas_payer_priv"]
-        self.multi_credentials = self._validate_multi_credentials(multi_credentials)
+        self.multi_credentials = self._validate_multi_credentials(
+            multi_credentials)
 
         # Initialize a new Job.
         if not escrow_addr and escrow_manifest:
-            self.factory_contract = self._init_factory(factory_addr, credentials)
+            self.factory_contract = self._init_factory(factory_addr,
+                                                       credentials)
             self._init_job(escrow_manifest)
 
         # Access an existing Job.
@@ -302,14 +302,15 @@ class Job:
 
         # Handle incorrect usage
         else:
-            raise ValueError("Job instantiation wrong, double-check arguments.")
+            raise ValueError(
+                "Job instantiation wrong, double-check arguments.")
 
     def launch(self, pub_key: bytes) -> bool:
         """Launches an escrow contract to the network, uploads the manifest
         to S3 with the public key of the Reputation Oracle and stores
         the S3 url to the escrow contract.
 
-        >>> from test_manifest import manifest
+        >>> from test.hmt_escrow.utils import manifest
         >>> credentials = {
         ... 	"gas_payer": "0x1413862C2B7054CDbfdc181B83962CB0FC11fD92",
         ... 	"gas_payer_priv": "28e516f1e2f99e96a48a23cea1f94ee5f073403a1c68e818263f0eb898f1c8e5"
@@ -374,7 +375,7 @@ class Job:
         """Sets the escrow contract to be ready to receive answers from the Recording Oracle.
         The contract needs to be deployed and funded first.
 
-        >>> from test_manifest import manifest
+        >>> from test.hmt_escrow.utils import manifest
         >>> credentials = {
         ... 	"gas_payer": "0x1413862C2B7054CDbfdc181B83962CB0FC11fD92",
         ... 	"gas_payer_priv": "28e516f1e2f99e96a48a23cea1f94ee5f073403a1c68e818263f0eb898f1c8e5"
@@ -414,8 +415,10 @@ class Job:
         recording_oracle_stake = int(
             Decimal(self.serialized_manifest["oracle_stake"]) * 100
         )
-        reputation_oracle = str(self.serialized_manifest["reputation_oracle_addr"])
-        recording_oracle = str(self.serialized_manifest["recording_oracle_addr"])
+        reputation_oracle = str(
+            self.serialized_manifest["reputation_oracle_addr"])
+        recording_oracle = str(
+            self.serialized_manifest["recording_oracle_addr"])
         hmt_amount = int(self.amount * 10 ** 18)
         hmtoken_contract = get_hmtoken()
 
@@ -473,7 +476,8 @@ class Job:
         ]
 
         try:
-            handle_transaction_with_retry(txn_func, self.retry, *func_args, **txn_info)
+            handle_transaction_with_retry(txn_func, self.retry, *func_args,
+                                          **txn_info)
             contract_is_setup = True
         except Exception as e:
             LOG.debug(
@@ -489,13 +493,15 @@ class Job:
             LOG.warning(f"{txn_event} failed with all credentials.")
 
         return (
-            str(self.status()) == str(Status.Pending) and self.balance() == hmt_amount
+                str(self.status()) == str(
+            Status.Pending) and self.balance() == hmt_amount
         )
 
-    def add_trusted_handlers(self, handlers: List[str], gas: int = GAS_LIMIT) -> bool:
+    def add_trusted_handlers(self, handlers: List[str],
+                             gas: int = GAS_LIMIT) -> bool:
         """Add trusted handlers that can freely transact with the contract and perform aborts and cancels for example.
 
-        >>> from test_manifest import manifest
+        >>> from test.hmt_escrow.utils import manifest
         >>> credentials = {
         ... 	"gas_payer": "0x1413862C2B7054CDbfdc181B83962CB0FC11fD92",
         ... 	"gas_payer_priv": "28e516f1e2f99e96a48a23cea1f94ee5f073403a1c68e818263f0eb898f1c8e5"
@@ -535,7 +541,8 @@ class Job:
         func_args = [handlers]
 
         try:
-            handle_transaction_with_retry(txn_func, self.retry, *func_args, **txn_info)
+            handle_transaction_with_retry(txn_func, self.retry, *func_args,
+                                          **txn_info)
             return True
         except Exception as e:
             LOG.info(
@@ -551,17 +558,17 @@ class Job:
         return trusted_handlers_added
 
     def bulk_payout(
-        self,
-        payouts: List[Tuple[str, Decimal]],
-        results: Dict,
-        pub_key: bytes,
-        gas: int = GAS_LIMIT,
+            self,
+            payouts: List[Tuple[str, Decimal]],
+            results: Dict,
+            pub_key: bytes,
+            gas: int = GAS_LIMIT,
     ) -> bool:
         """Performs a payout to multiple ethereum addresses. When the payout happens,
         final results are uploaded to IPFS and contract's state is updated to Partial or Paid
         depending on contract's balance.
 
-        >>> from test_manifest import manifest
+        >>> from test.hmt_escrow.utils import manifest
         >>> credentials = {
         ... 	"gas_payer": "0x1413862C2B7054CDbfdc181B83962CB0FC11fD92",
         ... 	"gas_payer_priv": "28e516f1e2f99e96a48a23cea1f94ee5f073403a1c68e818263f0eb898f1c8e5"
@@ -632,7 +639,8 @@ class Job:
 
         func_args = [eth_addrs, hmt_amounts, url, hash_, 1]
         try:
-            handle_transaction_with_retry(txn_func, self.retry, *func_args, **txn_info)
+            handle_transaction_with_retry(txn_func, self.retry, *func_args,
+                                          **txn_info)
             return self._bulk_paid() == True
         except Exception as e:
             LOG.debug(
@@ -664,7 +672,7 @@ class Job:
 
         The escrow contract is in Partial state after a partial bulk payout so it can be aborted.
 
-        >>> from test_manifest import manifest
+        >>> from test.hmt_escrow.utils import manifest
         >>> job = Job(credentials, manifest)
 
         The escrow contract is in Pending state after setup so it can be aborted.
@@ -733,7 +741,8 @@ class Job:
         }
 
         try:
-            handle_transaction_with_retry(txn_func, self.retry, *[], **txn_info)
+            handle_transaction_with_retry(txn_func, self.retry, *[],
+                                          **txn_info)
             # After abort the contract should be destroyed
             return w3.eth.getCode(self.job_contract.address) == b""
         except Exception as e:
@@ -741,7 +750,8 @@ class Job:
                 f"{txn_event} failed with main credentials: {self.gas_payer}, {self.gas_payer_priv} due to {e}. Using secondary ones..."
             )
 
-        job_aborted = self._raffle_txn(self.multi_credentials, txn_func, [], txn_event)
+        job_aborted = self._raffle_txn(self.multi_credentials, txn_func, [],
+                                       txn_event)
 
         if not job_aborted:
             LOG.exception(f"{txn_event} failed with all credentials.")
@@ -756,7 +766,7 @@ class Job:
         ... 	"gas_payer_priv": "28e516f1e2f99e96a48a23cea1f94ee5f073403a1c68e818263f0eb898f1c8e5"
         ... }
         >>> rep_oracle_pub_key = b"2dbc2c2c86052702e7c219339514b2e8bd4687ba1236c478ad41b43330b08488c12c8c1797aa181f3a4596a1bd8a0c18344ea44d6655f61fa73e56e743f79e0d"
-        >>> from test_manifest import manifest
+        >>> from test.hmt_escrow.utils import manifest
         >>> job = Job(credentials, manifest)
 
         The escrow contract is in Pending state after setup so it can be cancelled.
@@ -811,7 +821,8 @@ class Job:
         }
 
         try:
-            handle_transaction_with_retry(txn_func, self.retry, *[], **txn_info)
+            handle_transaction_with_retry(txn_func, self.retry, *[],
+                                          **txn_info)
             return self.status() == Status.Cancelled
         except Exception as e:
             LOG.info(
@@ -828,12 +839,12 @@ class Job:
         return self.status() == Status.Cancelled
 
     def store_intermediate_results(
-        self, results: Dict, pub_key: bytes, gas: int = GAS_LIMIT
+            self, results: Dict, pub_key: bytes, gas: int = GAS_LIMIT
     ) -> bool:
         """Recording Oracle stores intermediate results with Reputation Oracle's public key to S3
         and updates the contract's state.
 
-        >>> from test_manifest import manifest
+        >>> from test.hmt_escrow.utils import manifest
         >>> credentials = {
         ... 	"gas_payer": "0x1413862C2B7054CDbfdc181B83962CB0FC11fD92",
         ... 	"gas_payer_priv": "28e516f1e2f99e96a48a23cea1f94ee5f073403a1c68e818263f0eb898f1c8e5"
@@ -897,7 +908,8 @@ class Job:
         func_args = [url, hash_]
 
         try:
-            handle_transaction_with_retry(txn_func, self.retry, *func_args, **txn_info)
+            handle_transaction_with_retry(txn_func, self.retry, *func_args,
+                                          **txn_info)
             return True
         except Exception as e:
             LOG.info(
@@ -917,16 +929,16 @@ class Job:
         return results_stored
 
     def complete(
-        self,
-        gas: int = GAS_LIMIT,
-        blocking: bool = False,
-        retries: int = 3,
-        delay: int = 5,
-        backoff: int = 2,
+            self,
+            gas: int = GAS_LIMIT,
+            blocking: bool = False,
+            retries: int = 3,
+            delay: int = 5,
+            backoff: int = 2,
     ) -> bool:
         """Completes the Job if it has been paid.
 
-        >>> from test_manifest import manifest
+        >>> from test.hmt_escrow.utils import manifest
         >>> credentials = {
         ... 	"gas_payer": "0x1413862C2B7054CDbfdc181B83962CB0FC11fD92",
         ... 	"gas_payer_priv": "28e516f1e2f99e96a48a23cea1f94ee5f073403a1c68e818263f0eb898f1c8e5"
@@ -971,7 +983,8 @@ class Job:
         }
 
         try:
-            handle_transaction_with_retry(txn_func, self.retry, *[], **txn_info)
+            handle_transaction_with_retry(txn_func, self.retry, *[],
+                                          **txn_info)
             return self.status() == Status.Complete
         except Exception as e:
             LOG.info(
@@ -990,7 +1003,7 @@ class Job:
     def status(self, gas: int = GAS_LIMIT) -> Enum:
         """Returns the status of the Job.
 
-        >>> from test_manifest import manifest
+        >>> from test.hmt_escrow.utils import manifest
         >>> credentials = {
         ... 	"gas_payer": "0x1413862C2B7054CDbfdc181B83962CB0FC11fD92",
         ... 	"gas_payer_priv": "28e516f1e2f99e96a48a23cea1f94ee5f073403a1c68e818263f0eb898f1c8e5"
@@ -1014,7 +1027,7 @@ class Job:
     def balance(self, gas: int = GAS_LIMIT) -> int:
         """Retrieve the balance of a Job in HMT.
 
-        >>> from test_manifest import manifest
+        >>> from test.hmt_escrow.utils import manifest
         >>> credentials = {
         ... 	"gas_payer": "0x1413862C2B7054CDbfdc181B83962CB0FC11fD92",
         ... 	"gas_payer_priv": "28e516f1e2f99e96a48a23cea1f94ee5f073403a1c68e818263f0eb898f1c8e5"
@@ -1044,7 +1057,7 @@ class Job:
     def manifest(self, priv_key: bytes) -> Dict:
         """Retrieves the initial manifest used to setup a Job.
 
-        >>> from test_manifest import manifest
+        >>> from test.hmt_escrow.utils import manifest
         >>> credentials = {
         ... 	"gas_payer": "0x1413862C2B7054CDbfdc181B83962CB0FC11fD92",
         ... 	"gas_payer_priv": "28e516f1e2f99e96a48a23cea1f94ee5f073403a1c68e818263f0eb898f1c8e5"
@@ -1070,10 +1083,11 @@ class Job:
         """
         return download(self.manifest_url, priv_key)
 
-    def intermediate_results(self, priv_key: bytes, gas: int = GAS_LIMIT) -> Dict:
+    def intermediate_results(self, priv_key: bytes,
+                             gas: int = GAS_LIMIT) -> Dict:
         """Reputation Oracle retrieves the intermediate results stored by the Recording Oracle.
 
-        >>> from test_manifest import manifest
+        >>> from test.hmt_escrow.utils import manifest
         >>> credentials = {
         ... 	"gas_payer": "0x1413862C2B7054CDbfdc181B83962CB0FC11fD92",
         ... 	"gas_payer_priv": "28e516f1e2f99e96a48a23cea1f94ee5f073403a1c68e818263f0eb898f1c8e5"
@@ -1107,7 +1121,7 @@ class Job:
     def final_results(self, priv_key: bytes, gas: int = GAS_LIMIT) -> Dict:
         """Retrieves the final results stored by the Reputation Oracle.
 
-        >>> from test_manifest import manifest
+        >>> from test.hmt_escrow.utils import manifest
         >>> credentials = {
         ... 	"gas_payer": "0x1413862C2B7054CDbfdc181B83962CB0FC11fD92",
         ... 	"gas_payer_priv": "28e516f1e2f99e96a48a23cea1f94ee5f073403a1c68e818263f0eb898f1c8e5"
@@ -1182,13 +1196,13 @@ class Job:
         return Web3.toChecksumAddress(addr) == calculated_addr
 
     def _validate_multi_credentials(
-        self, multi_credentials: List[Tuple]
+            self, multi_credentials: List[Tuple]
     ) -> List[Tuple[Any, Any]]:
         """Validates whether the given ethereum private key maps to the address
         by calculating the checksum address from the private key and comparing that
         to the given address.
 
-        >>> from test_manifest import manifest
+        >>> from test.hmt_escrow.utils import manifest
         >>> credentials = {
         ...     "gas_payer": "0x1413862C2B7054CDbfdc181B83962CB0FC11fD92",
         ... 	"gas_payer_priv": "28e516f1e2f99e96a48a23cea1f94ee5f073403a1c68e818263f0eb898f1c8e5"
@@ -1222,14 +1236,14 @@ class Job:
         return valid_credentials
 
     def _validate_credentials(
-        self, multi_credentials: List[Tuple], **credentials
+            self, multi_credentials: List[Tuple], **credentials
     ) -> bool:
         """Validates whether the given ethereum private key maps to the address
         by calculating the checksum address from the private key and comparing that
         to the given address.
 
         Validating right credentials succeeds.
-        >>> from test_manifest import manifest
+        >>> from test.hmt_escrow.utils import manifest
         >>> credentials = {
         ...     "gas_payer": "0x1413862C2B7054CDbfdc181B83962CB0FC11fD92",
         ... 	"gas_payer_priv": "28e516f1e2f99e96a48a23cea1f94ee5f073403a1c68e818263f0eb898f1c8e5"
@@ -1262,11 +1276,11 @@ class Job:
         return self._eth_addr_valid(gas_payer_addr, gas_payer_priv)
 
     def _factory_contains_escrow(
-        self, escrow_addr: str, factory_addr: str, gas: int = GAS_LIMIT
+            self, escrow_addr: str, factory_addr: str, gas: int = GAS_LIMIT
     ) -> bool:
         """Checks whether a given factory address contains a given escrow address.
 
-        >>> from test_manifest import manifest
+        >>> from test.hmt_escrow.utils import manifest
         >>> credentials = {
         ... 	"gas_payer": "0x1413862C2B7054CDbfdc181B83962CB0FC11fD92",
         ... 	"gas_payer_priv": "28e516f1e2f99e96a48a23cea1f94ee5f073403a1c68e818263f0eb898f1c8e5",
@@ -1302,16 +1316,16 @@ class Job:
         )
 
     def _init_factory(
-        self,
-        factory_addr: Optional[str],
-        credentials: Dict[str, str],
-        gas: int = GAS_LIMIT,
+            self,
+            factory_addr: Optional[str],
+            credentials: Dict[str, str],
+            gas: int = GAS_LIMIT,
     ) -> Contract:
         """Takes an optional factory address and returns its contract representation. Alternatively
         a new factory is created.
 
         Initializing a new Job instance without a factory address succeeds.
-        >>> from test_manifest import manifest
+        >>> from test.hmt_escrow.utils import manifest
         >>> credentials = {
         ... 	"gas_payer": "0x1413862C2B7054CDbfdc181B83962CB0FC11fD92",
         ... 	"gas_payer_priv": "28e516f1e2f99e96a48a23cea1f94ee5f073403a1c68e818263f0eb898f1c8e5"
@@ -1351,7 +1365,7 @@ class Job:
     def _bulk_paid(self, gas: int = GAS_LIMIT) -> int:
         """Checks if the last bulk payment has succeeded.
 
-        >>> from test_manifest import manifest
+        >>> from test.hmt_escrow.utils import manifest
         >>> credentials = {
         ... 	"gas_payer": "0x1413862C2B7054CDbfdc181B83962CB0FC11fD92",
         ... 	"gas_payer_priv": "28e516f1e2f99e96a48a23cea1f94ee5f073403a1c68e818263f0eb898f1c8e5"
@@ -1388,7 +1402,7 @@ class Job:
     def _last_escrow_addr(self, gas: int = GAS_LIMIT) -> str:
         """Gets the last deployed escrow contract address of the initialized factory contract.
 
-        >>> from test_manifest import manifest
+        >>> from test.hmt_escrow.utils import manifest
         >>> credentials = {
         ... 	"gas_payer": "0x1413862C2B7054CDbfdc181B83962CB0FC11fD92",
         ... 	"gas_payer_priv": "28e516f1e2f99e96a48a23cea1f94ee5f073403a1c68e818263f0eb898f1c8e5"
@@ -1412,10 +1426,11 @@ class Job:
             {"from": self.gas_payer, "gas": Wei(gas)}
         )
 
-    def _create_escrow(self, trusted_handlers=[], gas: int = GAS_LIMIT) -> bool:
+    def _create_escrow(self, trusted_handlers=[],
+                       gas: int = GAS_LIMIT) -> bool:
         """Launches a new escrow contract to the ethereum network.
 
-        >>> from test_manifest import manifest
+        >>> from test.hmt_escrow.utils import manifest
         >>> multi_credentials = [("0x61F9F0B31eacB420553da8BCC59DC617279731Ac", "486a0621e595dd7fcbe5608cbbeec8f5a8b5cabe7637f11eccfc7acd408c3a0e"), ("0x6b7E3C31F34cF38d1DFC1D9A8A59482028395809", "f22d4fc42da79aa5ba839998a0a9f2c2c45f5e55ee7f1504e464d2c71ca199e1")]
         >>> trusted_handlers = [addr for addr, priv_key in multi_credentials]
         >>> credentials = {
@@ -1454,7 +1469,8 @@ class Job:
         func_args = [trusted_handlers]
 
         try:
-            handle_transaction_with_retry(txn_func, self.retry, *func_args, **txn_info)
+            handle_transaction_with_retry(txn_func, self.retry, *func_args,
+                                          **txn_info)
             return True
         except Exception as e:
             LOG.info(
@@ -1471,7 +1487,8 @@ class Job:
         return escrow_created
 
     def _raffle_txn(
-        self, multi_creds, txn_func, txn_args, txn_event, gas: int = GAS_LIMIT
+            self, multi_creds, txn_func, txn_args, txn_event,
+            gas: int = GAS_LIMIT
     ):
         """Takes in multiple credentials, loops through each and performs the given transaction.
 
@@ -1508,345 +1525,3 @@ class Job:
                 )
 
         return txn_succeeded
-
-
-class JobTestCase(unittest.TestCase):
-    def setUp(self):
-        self.credentials = {
-            "gas_payer": "0x1413862C2B7054CDbfdc181B83962CB0FC11fD92",
-            "gas_payer_priv": "28e516f1e2f99e96a48a23cea1f94ee5f073403a1c68e818263f0eb898f1c8e5",
-        }
-        self.rep_oracle_pub_key = b"2dbc2c2c86052702e7c219339514b2e8bd4687ba1236c478ad41b43330b08488c12c8c1797aa181f3a4596a1bd8a0c18344ea44d6655f61fa73e56e743f79e0d"
-        self.job = Job(self.credentials, manifest)
-
-    def test_status(self):
-        self.assertTrue(self.job.launch(self.rep_oracle_pub_key))
-        self.assertEqual(status(self.job.job_contract, self.job.gas_payer), Status(1))
-
-    def test_manifest_url(self):
-        self.assertTrue(self.job.launch(self.rep_oracle_pub_key))
-        self.assertTrue(self.job.setup())
-        self.assertEqual(
-            manifest_hash(self.job.job_contract, self.job.gas_payer),
-            self.job.manifest_hash,
-        )
-
-    def test_job_init(self):
-
-        # Creating a new Job instance initializes the critical attributes correctly.
-        self.assertEqual(self.job.gas_payer, self.credentials["gas_payer"])
-        self.assertEqual(self.job.gas_payer_priv, self.credentials["gas_payer_priv"])
-        self.assertEqual(self.job.serialized_manifest["oracle_stake"], "0.05")
-        self.assertEqual(self.job.amount, Decimal("100.0"))
-
-        # Initializing a new Job instance with a factory address succeeds.
-        factory_addr = deploy_factory(**(self.credentials))
-        self.job = Job(self.credentials, manifest, factory_addr)
-        self.assertTrue(self.job.factory_contract.address, factory_addr)
-        self.assertTrue(self.job.launch(self.rep_oracle_pub_key))
-        self.assertTrue(self.job.setup())
-        self.assertTrue(
-            launcher(self.job.job_contract, self.credentials["gas_payer"]).lower(),
-            self.job.factory_contract.address.lower(),
-        )
-
-        # Initializing an existing Job instance with a factory and escrow address succeeds.
-        self.credentials[
-            "rep_oracle_priv_key"
-        ] = b"28e516f1e2f99e96a48a23cea1f94ee5f073403a1c68e818263f0eb898f1c8e5"
-        escrow_addr = self.job.job_contract.address
-        factory_addr = self.job.factory_contract.address
-        manifest_url = self.job.manifest_url
-        new_job = Job(
-            credentials=self.credentials,
-            factory_addr=factory_addr,
-            escrow_addr=escrow_addr,
-        )
-        self.assertEqual(new_job.manifest_url, manifest_url)
-        self.assertEqual(new_job.job_contract.address, escrow_addr)
-        self.assertEqual(new_job.factory_contract.address, factory_addr)
-        with self.assertRaises(AttributeError):
-            new_job.launch(self.rep_oracle_pub_key)
-
-    def test_job_launch(self):
-        self.assertTrue(self.job.launch(self.rep_oracle_pub_key))
-        self.assertEqual(self.job.status(), Status(1))
-        multi_credentials = [
-            (
-                "0x61F9F0B31eacB420553da8BCC59DC617279731Ac",
-                "486a0621e595dd7fcbe5608cbbeec8f5a8b5cabe7637f11eccfc7acd408c3a0e",
-            ),
-            (
-                "0x6b7E3C31F34cF38d1DFC1D9A8A59482028395809",
-                "f22d4fc42da79aa5ba839998a0a9f2c2c45f5e55ee7f1504e464d2c71ca199e1",
-            ),
-        ]
-        self.job = Job(self.credentials, manifest, multi_credentials=multi_credentials)
-
-        # Inject wrong credentials on purpose to test out raffling
-
-        self.job.gas_payer_priv = (
-            "657b6497a355a3982928d5515d48a84870f057c4d16923eb1d104c0afada9aa8"
-        )
-        self.job.multi_credentials = [
-            (
-                "0x61F9F0B31eacB420553da8BCC59DC617279731Ac",
-                "28e516f1e2f99e96a48a23cea1f94ee5f073403a1c68e818263f0eb898f1c8e5",
-            ),
-            (
-                "0x6b7E3C31F34cF38d1DFC1D9A8A59482028395809",
-                "f22d4fc42da79aa5ba839998a0a9f2c2c45f5e55ee7f1504e464d2c71ca199e1",
-            ),
-        ]
-        self.assertTrue(self.job.launch(self.rep_oracle_pub_key))
-        self.assertEqual(self.job.status(), Status(1))
-
-        # Make sure we launched with raffled credentials
-
-        self.assertEqual(
-            self.job.gas_payer, "0x6b7E3C31F34cF38d1DFC1D9A8A59482028395809"
-        )
-        self.assertEqual(
-            self.job.gas_payer_priv,
-            "f22d4fc42da79aa5ba839998a0a9f2c2c45f5e55ee7f1504e464d2c71ca199e1",
-        )
-
-    def test_job_setup(self):
-
-        # A Job can't be setup without deploying it first.
-
-        self.assertFalse(self.job.setup())
-        multi_credentials = [
-            (
-                "0x61F9F0B31eacB420553da8BCC59DC617279731Ac",
-                "28e516f1e2f99e96a48a23cea1f94ee5f073403a1c68e818263f0eb898f1c8e5",
-            ),
-            (
-                "0x6b7E3C31F34cF38d1DFC1D9A8A59482028395809",
-                "f22d4fc42da79aa5ba839998a0a9f2c2c45f5e55ee7f1504e464d2c71ca199e1",
-            ),
-        ]
-        self.job = Job(self.credentials, manifest, multi_credentials=multi_credentials)
-        self.assertTrue(self.job.launch(self.rep_oracle_pub_key))
-        self.assertTrue(self.job.setup())
-
-    def test_job_add_trusted_handlers(self):
-
-        # Make sure we se set our gas payer as a trusted handler by default.
-
-        self.assertTrue(self.job.launch(self.rep_oracle_pub_key))
-        self.assertTrue(
-            is_trusted_handler(
-                self.job.job_contract, self.job.gas_payer, self.job.gas_payer
-            )
-        )
-        trusted_handlers = [
-            "0x61F9F0B31eacB420553da8BCC59DC617279731Ac",
-            "0xD979105297fB0eee83F7433fC09279cb5B94fFC6",
-        ]
-        self.assertTrue(self.job.add_trusted_handlers(trusted_handlers))
-        self.assertTrue(
-            is_trusted_handler(
-                self.job.job_contract,
-                "0x61F9F0B31eacB420553da8BCC59DC617279731Ac",
-                self.job.gas_payer,
-            )
-        )
-        self.assertTrue(
-            is_trusted_handler(
-                self.job.job_contract,
-                "0xD979105297fB0eee83F7433fC09279cb5B94fFC6",
-                self.job.gas_payer,
-            )
-        )
-
-    def test_job_bulk_payout(self):
-        self.assertTrue(self.job.launch(self.rep_oracle_pub_key))
-        self.assertTrue(self.job.setup())
-        payouts = [
-            ("0x6b7E3C31F34cF38d1DFC1D9A8A59482028395809", Decimal("20.0")),
-            ("0x852023fbb19050B8291a335E5A83Ac9701E7B4E6", Decimal("50.0")),
-        ]
-        self.assertTrue(self.job.bulk_payout(payouts, {}, self.rep_oracle_pub_key))
-
-        # The escrow contract is still in Partial state as there's still balance left.
-
-        self.assertEqual(self.job.balance(), 30000000000000000000)
-        self.assertEqual(self.job.status(), Status(3))
-
-        # Trying to pay more than the contract balance results in failure.
-
-        payouts = [("0x9d689b8f50Fd2CAec716Cc5220bEd66E03F07B5f", Decimal("40.0"))]
-        self.assertFalse(self.job.bulk_payout(payouts, {}, self.rep_oracle_pub_key))
-
-        # Paying the remaining amount empties the escrow and updates the status correctly.
-
-        payouts = [("0x9d689b8f50Fd2CAec716Cc5220bEd66E03F07B5f", Decimal("30.0"))]
-        self.assertTrue(self.job.bulk_payout(payouts, {}, self.rep_oracle_pub_key))
-        self.assertEqual(self.job.balance(), 0)
-        self.assertEqual(self.job.status(), Status(4))
-
-        multi_credentials = [
-            (
-                "0x61F9F0B31eacB420553da8BCC59DC617279731Ac",
-                "486a0621e595dd7fcbe5608cbbeec8f5a8b5cabe7637f11eccfc7acd408c3a0e",
-            ),
-            (
-                "0x6b7E3C31F34cF38d1DFC1D9A8A59482028395809",
-                "f22d4fc42da79aa5ba839998a0a9f2c2c45f5e55ee7f1504e464d2c71ca199e1",
-            ),
-        ]
-        self.job = Job(self.credentials, manifest, multi_credentials=multi_credentials)
-        self.assertTrue(self.job.launch(self.rep_oracle_pub_key))
-        self.assertTrue(self.job.setup())
-        payouts = [
-            ("0x6b7E3C31F34cF38d1DFC1D9A8A59482028395809", Decimal("20.0")),
-            ("0x852023fbb19050B8291a335E5A83Ac9701E7B4E6", Decimal("50.0")),
-        ]
-        self.assertTrue(self.job.bulk_payout(payouts, {}, self.rep_oracle_pub_key))
-
-    def test_job_abort(self):
-
-        # The escrow contract is in Paid state after the full bulk payout and it can't be aborted.
-
-        self.assertTrue(self.job.launch(self.rep_oracle_pub_key))
-        self.assertTrue(self.job.setup())
-        payouts = [("0x852023fbb19050B8291a335E5A83Ac9701E7B4E6", Decimal("100.0"))]
-        self.assertTrue(
-            self.job.bulk_payout(payouts, {"results": 0}, self.rep_oracle_pub_key)
-        )
-        self.assertFalse(self.job.abort())
-        self.assertEqual(self.job.status(), Status(4))
-
-        # Trusted handler should be able to abort an existing contract
-
-        self.job = Job(self.credentials, manifest)
-        self.assertTrue(self.job.launch(self.rep_oracle_pub_key))
-        self.assertTrue(self.job.setup())
-        trusted_handler = "0x6b7E3C31F34cF38d1DFC1D9A8A59482028395809"
-        self.assertTrue(self.job.add_trusted_handlers([trusted_handler]))
-
-        handler_credentials = {
-            "gas_payer": "0x6b7E3C31F34cF38d1DFC1D9A8A59482028395809",
-            "gas_payer_priv": "f22d4fc42da79aa5ba839998a0a9f2c2c45f5e55ee7f1504e464d2c71ca199e1",
-            "rep_oracle_priv_key": b"28e516f1e2f99e96a48a23cea1f94ee5f073403a1c68e818263f0eb898f1c8e5",
-        }
-        access_job = Job(
-            credentials=handler_credentials,
-            factory_addr=self.job.factory_contract.address,
-            escrow_addr=self.job.job_contract.address,
-        )
-        self.assertTrue(access_job.abort())
-
-    def test_job_cancel(self):
-        self.assertTrue(self.job.launch(self.rep_oracle_pub_key))
-        self.assertTrue(self.job.setup())
-        payouts = [("0x6b7E3C31F34cF38d1DFC1D9A8A59482028395809", Decimal("20.0"))]
-        self.assertTrue(self.job.bulk_payout(payouts, {}, self.rep_oracle_pub_key))
-        self.assertEqual(self.job.status(), Status(3))
-
-        # The escrow contract is in Paid state after the second payout and it can't be cancelled.
-
-        payouts = [("0x852023fbb19050B8291a335E5A83Ac9701E7B4E6", Decimal("80.0"))]
-        self.assertTrue(
-            self.job.bulk_payout(payouts, {"results": 0}, self.rep_oracle_pub_key)
-        )
-        self.assertFalse(self.job.cancel())
-        self.assertEqual(self.job.status(), Status(4))
-
-    def test_job_status(self):
-        self.assertTrue(self.job.launch(self.rep_oracle_pub_key))
-        self.assertEqual(self.job.status(), Status(1))
-
-    def test_job_balance(self):
-        self.assertTrue(self.job.launch(self.rep_oracle_pub_key))
-        self.assertTrue(self.job.setup())
-        self.assertEqual(self.job.balance(), 100000000000000000000)
-
-    def test_launch_failure(self):
-        """ Test _launch raises error on failure to create contract """
-
-        handler_mock = MagicMock()
-        handler_mock.side_effect = Exception("")
-
-        e = None
-        with self.assertRaises(Exception) as e:
-            with patch("hmt_escrow.eth_bridge.handle_transaction", handler_mock):
-                self.job.launch(b"")
-
-        self.assertIsNotNone(e)
-        self.assertEqual(str(e.exception), "Unable to create escrow")
-
-    def test__raffle_txn_retry(self):
-        """ Test general retry logic """
-
-        retries = 4
-        delay = 0.01
-        backoff = 2
-        self.job.retry = Retry(retries=retries, delay=delay, backoff=backoff)
-
-        txn_mock = MagicMock()
-        handler_mock = MagicMock(side_effect=Exception)
-        sleep_mock = MagicMock()
-
-        with patch("hmt_escrow.eth_bridge.handle_transaction", handler_mock), patch(
-            "hmt_escrow.eth_bridge.sleep", sleep_mock
-        ):
-            success = self.job._raffle_txn(
-                multi_creds=[("1", "11")],
-                txn_func=txn_mock,
-                txn_args=[],
-                txn_event="Transfer",
-            )
-
-            self.assertFalse(success)
-
-            self.assertEqual(
-                handler_mock.call_args_list,
-                [
-                    call(txn_mock, gas_payer="1", gas_payer_priv="11", gas=6700000)
-                    for i in range(5)
-                ],
-            )
-            self.assertEqual(
-                sleep_mock.call_args_list,
-                [call(delay * backoff ** i) for i in range(retries)],
-            )
-
-            sleep_mock.reset_mock()
-            handler_mock.reset_mock()
-            handler_mock.side_effect = Exception
-
-            # no retries
-            self.job.retry = Retry()
-
-            success = self.job._raffle_txn(
-                multi_creds=[("1", "11")],
-                txn_func=txn_mock,
-                txn_args=[],
-                txn_event="Transfer",
-            )
-
-            self.assertFalse(success)
-            self.assertEqual(
-                handler_mock.call_args_list,
-                [call(txn_mock, gas_payer="1", gas_payer_priv="11", gas=6700000)],
-            )
-            self.assertEqual(sleep_mock.call_args_list, [])
-
-    def test_get_hmt_balance(self):
-        """ Test wallet HMT balance is OK """
-        self.assertGreater(
-            utils.get_hmt_balance(
-                "0x1413862C2B7054CDbfdc181B83962CB0FC11fD92",
-                "0x56B532F1D090E4edb1c92F30d3087771AE6B6992",
-                get_w3(),
-            ),
-            1000000,
-        )
-
-
-if __name__ == "__main__":
-    from test_manifest import manifest
-
-    unittest.main(exit=True)
